v2024.1.3 Switched to fractional quantities/volumes display; some more fixes to schematic-related costs

+++ DU-Industry-Tool/Classes/Recipe.cs
@@ -264,7 +264,7 @@ namespace DU_Industry_Tool
         [JsonIgnore]
         public string SchemaType { get; set; }
         [JsonIgnore]
-        public int SchemaQty { get; set; }
+        public decimal SchemaQty { get; set; }
         [JsonIgnore]
         public decimal SchemaAmt { get; set; }
         [JsonIgnore]
